+++ .tmux/plugins/tmux-resurrect/scripts/process_restore_helpers.sh
@@ -13,6 +13,7 @@ restore_pane_process() {
 	local window_number="$3"
 	local pane_index="$4"
 	local dir="$5"
+	local command
 	if _process_should_be_restored "$pane_full_command" "$session_name" "$window_number" "$pane_index"; then
 		tmux switch-client -t "${session_name}:${window_number}"
 		tmux select-pane -t "$pane_index"
@@ -25,15 +26,16 @@ restore_pane_process() {
 				local strategy_file="$(_get_strategy_file "$inline_strategy")"
 				local inline_strategy="$($strategy_file "$pane_full_command" "$dir")"
 			fi
-			tmux send-keys "$inline_strategy" "C-m"
+			command="$inline_strategy"
 		elif _strategy_exists "$pane_full_command"; then
 			local strategy_file="$(_get_strategy_file "$pane_full_command")"
 			local strategy_command="$($strategy_file "$pane_full_command" "$dir")"
-			tmux send-keys "$strategy_command" "C-m"
+			command="$strategy_command"
 		else
-			# just invoke the command
-			tmux send-keys "$pane_full_command" "C-m"
+			# just invoke the raw command
+			command="$pane_full_command"
 		fi
+		tmux send-keys -t "${session_name}:${window_number}.${pane_index}" "$command" "C-m"
 	fi
 }
 
@@ -110,6 +112,32 @@ _get_proc_restore_element() {
 	echo "$1" | sed "s/.*${inline_strategy_token}//"
 }
 
+# given full command: 'ruby /Users/john/bin/my_program arg1 arg2'
+# and inline strategy: '~bin/my_program->my_program *'
+# returns: 'arg1 arg2'
+_get_command_arguments() {
+	local pane_full_command="$1"
+	local match="$2"
+	if _proc_starts_with_tildae "$match"; then
+		match="$(remove_first_char "$match")"
+	fi
+	echo "$pane_full_command" | sed "s,^.*${match}[^ ]* *,,"
+}
+
+_get_proc_restore_command() {
+	local pane_full_command="$1"
+	local proc="$2"
+	local match="$3"
+	local restore_element="$(_get_proc_restore_element "$proc")"
+	if [[ "$restore_element" =~ " ${inline_strategy_arguments_token}" ]]; then
+		# replaces "%" with command arguments
+		local command_arguments="$(_get_command_arguments "$pane_full_command" "$match")"
+		echo "$restore_element" | sed "s,${inline_strategy_arguments_token},${command_arguments},"
+	else
+		echo "$restore_element"
+	fi
+}
+
 _restore_list() {
 	local user_processes="$(get_tmux_option "$restore_processes_option" "$restore_processes")"
 	local default_processes="$(get_tmux_option "$default_proc_list_option" "$default_proc_list")"
@@ -135,7 +163,7 @@ _get_inline_strategy() {
 		if [[ "$proc" =~ "$inline_strategy_token" ]]; then
 			match="$(_get_proc_match_element "$proc")"
 			if _proc_matches_full_command "$pane_full_command" "$match"; then
-				echo "$(_get_proc_restore_element "$proc")"
+				echo "$(_get_proc_restore_command "$pane_full_command" "$proc" "$match")"
 			fi
 		fi
 	done