+++ .tmux/plugins/tmux-resurrect/scripts/restore.sh
@@ -76,6 +76,14 @@ is_restoring_pane_contents() {
 	[ "$RESTORE_PANE_CONTENTS" == "true" ]
 }
 
+restored_session_0_true() {
+	RESTORED_SESSION_0="true"
+}
+
+has_restored_session_0() {
+	[ "$RESTORED_SESSION_0" == "true" ]
+}
+
 window_exists() {
 	local session_name="$1"
 	local window_number="$2"
@@ -100,7 +108,11 @@ tmux_socket() {
 # tmux server each time.
 cache_tmux_default_command() {
 	local default_shell="$(get_tmux_option "default-shell" "")"
-	export TMUX_DEFAULT_COMMAND="$(get_tmux_option "default-command" "$default_shell")"
+	local opt=""
+	if [ "$(basename "$default_shell")" == "bash" ]; then
+		opt="-l "
+	fi
+	export TMUX_DEFAULT_COMMAND="$(get_tmux_option "default-command" "$opt$default_shell")"
 }
 
 tmux_default_command() {
@@ -108,36 +120,35 @@ tmux_default_command() {
 }
 
 pane_creation_command() {
-	echo "cat '$(pane_contents_file "${1}:${2}.${3}")'; exec $(tmux_default_command)"
+	echo "cat '$(pane_contents_file "restore" "${1}:${2}.${3}")'; exec $(tmux_default_command)"
 }
 
 new_window() {
 	local session_name="$1"
 	local window_number="$2"
-	local window_name="$3"
-	local dir="$4"
-	local pane_index="$5"
+	local dir="$3"
+	local pane_index="$4"
 	local pane_id="${session_name}:${window_number}.${pane_index}"
+	dir="${dir/#\~/$HOME}"
 	if is_restoring_pane_contents && pane_contents_file_exists "$pane_id"; then
 		local pane_creation_command="$(pane_creation_command "$session_name" "$window_number" "$pane_index")"
-		tmux new-window -d -t "${session_name}:${window_number}" -n "$window_name" -c "$dir" "$pane_creation_command"
+		tmux new-window -d -t "${session_name}:${window_number}" -c "$dir" "$pane_creation_command"
 	else
-		tmux new-window -d -t "${session_name}:${window_number}" -n "$window_name" -c "$dir"
+		tmux new-window -d -t "${session_name}:${window_number}" -c "$dir"
 	fi
 }
 
 new_session() {
 	local session_name="$1"
 	local window_number="$2"
-	local window_name="$3"
-	local dir="$4"
-	local pane_index="$5"
+	local dir="$3"
+	local pane_index="$4"
 	local pane_id="${session_name}:${window_number}.${pane_index}"
 	if is_restoring_pane_contents && pane_contents_file_exists "$pane_id"; then
 		local pane_creation_command="$(pane_creation_command "$session_name" "$window_number" "$pane_index")"
-		TMUX="" tmux -S "$(tmux_socket)" new-session -d -s "$session_name" -n "$window_name" -c "$dir" "$pane_creation_command"
+		TMUX="" tmux -S "$(tmux_socket)" new-session -d -s "$session_name" -c "$dir" "$pane_creation_command"
 	else
-		TMUX="" tmux -S "$(tmux_socket)" new-session -d -s "$session_name" -n "$window_name" -c "$dir"
+		TMUX="" tmux -S "$(tmux_socket)" new-session -d -s "$session_name" -c "$dir"
 	fi
 	# change first window number if necessary
 	local created_window_num="$(first_window_num)"
@@ -149,9 +160,8 @@ new_session() {
 new_pane() {
 	local session_name="$1"
 	local window_number="$2"
-	local window_name="$3"
-	local dir="$4"
-	local pane_index="$5"
+	local dir="$3"
+	local pane_index="$4"
 	local pane_id="${session_name}:${window_number}.${pane_index}"
 	if is_restoring_pane_contents && pane_contents_file_exists "$pane_id"; then
 		local pane_creation_command="$(pane_creation_command "$session_name" "$window_number" "$pane_index")"
@@ -160,21 +170,23 @@ new_pane() {
 		tmux split-window -t "${session_name}:${window_number}" -c "$dir"
 	fi
 	# minimize window so more panes can fit
-	tmux resize-pane  -t "${session_name}:${window_number}" -U "999"
+	tmux resize-pane -t "${session_name}:${window_number}" -U "999"
 }
 
 restore_pane() {
 	local pane="$1"
-	while IFS=$d read line_type session_name window_number window_name window_active window_flags pane_index dir pane_active pane_command pane_full_command; do
+	while IFS=$d read line_type session_name window_number window_active window_flags pane_index pane_title dir pane_active pane_command pane_full_command; do
 		dir="$(remove_first_char "$dir")"
-		window_name="$(remove_first_char "$window_name")"
 		pane_full_command="$(remove_first_char "$pane_full_command")"
+		if [ "$session_name" == "0" ]; then
+			restored_session_0_true
+		fi
 		if pane_exists "$session_name" "$window_number" "$pane_index"; then
 			if is_restoring_from_scratch; then
 				# overwrite the pane
 				# happens only for the first pane if it's the only registered pane for the whole tmux server
 				local pane_id="$(tmux display-message -p -F "#{pane_id}" -t "$session_name:$window_number")"
-				new_pane "$session_name" "$window_number" "$window_name" "$dir" "$pane_index"
+				new_pane "$session_name" "$window_number" "$dir" "$pane_index"
 				tmux kill-pane -t "$pane_id"
 			else
 				# Pane exists, no need to create it!
@@ -182,12 +194,14 @@ restore_pane() {
 				register_existing_pane "$session_name" "$window_number" "$pane_index"
 			fi
 		elif window_exists "$session_name" "$window_number"; then
-			new_pane "$session_name" "$window_number" "$window_name" "$dir" "$pane_index"
+			new_pane "$session_name" "$window_number" "$dir" "$pane_index"
 		elif session_exists "$session_name"; then
-			new_window "$session_name" "$window_number" "$window_name" "$dir" "$pane_index"
+			new_window "$session_name" "$window_number" "$dir" "$pane_index"
 		else
-			new_session "$session_name" "$window_number" "$window_name" "$dir" "$pane_index"
+			new_session "$session_name" "$window_number" "$dir" "$pane_index"
 		fi
+		# set pane title
+		tmux select-pane -t "$session_name:$window_number.$pane_index" -T "$pane_title"
 	done < <(echo "$pane")
 }
 
@@ -258,29 +272,33 @@ restore_all_panes() {
 			restore_pane "$line"
 		fi
 	done < $(last_resurrect_file)
-	if is_restoring_pane_contents; then
-		pane_content_files_cleanup
+}
+
+handle_session_0() {
+	if is_restoring_from_scratch && ! has_restored_session_0; then
+		local current_session="$(tmux display -p "#{client_session}")"
+		if [ "$current_session" == "0" ]; then
+			tmux switch-client -n
+		fi
+		tmux kill-session -t "0"
 	fi
 }
 
-restore_pane_layout_for_each_window() {
+restore_window_properties() {
+	local window_name
 	\grep '^window' $(last_resurrect_file) |
-		while IFS=$d read line_type session_name window_number window_active window_flags window_layout; do
+		while IFS=$d read line_type session_name window_number window_name window_active window_flags window_layout automatic_rename; do
 			tmux select-layout -t "${session_name}:${window_number}" "$window_layout"
-		done
-}
 
-restore_shell_history() {
-	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ { print $2, $3, $7, $10; }' $(last_resurrect_file) |
-		while IFS=$d read session_name window_number pane_index pane_command; do
-			if ! is_pane_registered_as_existing "$session_name" "$window_number" "$pane_index"; then
-				if [ "$pane_command" == "bash" ]; then
-					local pane_id="$session_name:$window_number.$pane_index"
-					# tmux send-keys has -R option that should reset the terminal.
-					# However, appending 'clear' to the command seems to work more reliably.
-					local read_command="history -r '$(resurrect_history_file "$pane_id")'; clear"
-					tmux send-keys -t "$pane_id" "$read_command" C-m
-				fi
+			# Below steps are properly handling window names and automatic-rename
+			# option. `rename-window` is an extra command in some scenarios, but we
+			# opted for always doing it to keep the code simple.
+			window_name="$(remove_first_char "$window_name")"
+			tmux rename-window -t "${session_name}:${window_number}" "$window_name"
+			if [ "${automatic_rename}" = ":" ]; then
+				tmux set-option -u -t "${session_name}:${window_number}" automatic-rename
+			else
+				tmux set-option -t "${session_name}:${window_number}" automatic-rename "$automatic_rename"
 			fi
 		done
 }
@@ -288,8 +306,8 @@ restore_shell_history() {
 restore_all_pane_processes() {
 	if restore_pane_processes_enabled; then
 		local pane_full_command
-		awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $11 !~ "^:$" { print $2, $3, $7, $8, $11; }' $(last_resurrect_file) |
-			while IFS=$d read session_name window_number pane_index dir pane_full_command; do
+		awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $11 !~ "^:$" { print $2, $3, $6, $8, $11; }' $(last_resurrect_file) |
+			while IFS=$d read -r session_name window_number pane_index dir pane_full_command; do
 				dir="$(remove_first_char "$dir")"
 				pane_full_command="$(remove_first_char "$pane_full_command")"
 				restore_pane_process "$pane_full_command" "$session_name" "$window_number" "$pane_index" "$dir"
@@ -298,7 +316,7 @@ restore_all_pane_processes() {
 }
 
 restore_active_pane_for_each_window() {
-	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $9 == 1 { print $2, $3, $7; }' $(last_resurrect_file) |
+	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $9 == 1 { print $2, $3, $6; }' $(last_resurrect_file) |
 		while IFS=$d read session_name window_number active_pane; do
 			tmux switch-client -t "${session_name}:${window_number}"
 			tmux select-pane -t "$active_pane"
@@ -306,7 +324,7 @@ restore_active_pane_for_each_window() {
 }
 
 restore_zoomed_windows() {
-	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $6 ~ /Z/ && $9 == 1 { print $2, $3; }' $(last_resurrect_file) |
+	awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $5 ~ /Z/ && $9 == 1 { print $2, $3; }' $(last_resurrect_file) |
 		while IFS=$d read session_name window_number; do
 			tmux resize-pane -t "${session_name}:${window_number}" -Z
 		done
@@ -322,7 +340,7 @@ restore_grouped_sessions() {
 }
 
 restore_active_and_alternate_windows() {
-	awk 'BEGIN { FS="\t"; OFS="\t" } /^window/ && $5 ~ /[*-]/ { print $2, $4, $3; }' $(last_resurrect_file) |
+	awk 'BEGIN { FS="\t"; OFS="\t" } /^window/ && $6 ~ /[*-]/ { print $2, $5, $3; }' $(last_resurrect_file) |
 		sort -u |
 		while IFS=$d read session_name active_window window_number; do
 			tmux switch-client -t "${session_name}:${window_number}"
@@ -337,14 +355,22 @@ restore_active_and_alternate_sessions() {
 	done < $(last_resurrect_file)
 }
 
+# A cleanup that happens after 'restore_all_panes' seems to fix fish shell
+# users' restore problems.
+cleanup_restored_pane_contents() {
+	if is_restoring_pane_contents; then
+		rm "$(pane_contents_dir "restore")"/*
+	fi
+}
+
 main() {
 	if supported_tmux_version_ok && check_saved_session_exists; then
 		start_spinner "Restoring..." "Tmux restore complete!"
+		execute_hook "pre-restore-all"
 		restore_all_panes
-		restore_pane_layout_for_each_window >/dev/null 2>&1
-		if save_bash_history_option_on; then
-			restore_shell_history
-		fi
+		handle_session_0
+		restore_window_properties >/dev/null 2>&1
+		execute_hook "pre-restore-pane-processes"
 		restore_all_pane_processes
 		# below functions restore exact cursor positions
 		restore_active_pane_for_each_window
@@ -352,6 +378,8 @@ main() {
 		restore_grouped_sessions  # also restores active and alt windows for grouped sessions
 		restore_active_and_alternate_windows
 		restore_active_and_alternate_sessions
+		cleanup_restored_pane_contents
+		execute_hook "post-restore-all"
 		stop_spinner
 		display_message "Tmux restore complete!"
 	fi