+++ .tmux/plugins/tmux-resurrect/lib/tmux-test/CHANGELOG.md
@@ -8,6 +8,14 @@
 - remove tmux-test repo as a submodule from self, this causes issues with
   `$ git submodule update --recursive --init` command that some users use for
   managing other plugins
+- add new helper `teardown_helper`
+- add `run_tests` helper
+- change CLI syntax for choosing vagrant machine to run the tests on
+- enable running just a single test via `run_tests` cli interface
+- add `--keep-running` cli option to continue running vagrant after the tests
+  are done executing
+- start using tmux 2.0 for tests
+- use tmux 2.5 for tests
 
 ### v0.2.0, 2015-02-22
 - `setup` script gitignores `tests/helpers.sh`